update installer to not use root as user

+++ scripts/install.sh
@@ -1,19 +1,33 @@
 #!/bin/bash
 
 ###
-# Install elmasy.tar on a server
+# Install elmasy.tar on a Debian server
 ###
 
-if [ $1 == "" ]
+
+# Remove leftover file, if error occured before
+if [ -e elmasy.tar ]
 then
-    echo "Usage: ./$0 /path/to/elmasy.tar"
-    exit
-elif [[ $(file $1) != *"tar archive"* ]]
+    rm elmasy.tar
+fi
+
+# Create "elmasy" user, if not created before 
+if [ $(id -u elmasy > /dev/null 2>&1; echo $?) != 0 ]
 then
-    echo "$1 is not a TAR file!"
+    echo "Creating elmasy user..."
+    adduser --system --no-create-home --gecos "" --disabled-login elmasy
+fi
+
+
+OUTPUT=$(wget "https://github.com/elmasy-com/elmasy/raw/main/build/elmasy.tar" 2>&1)
+if [ $? != 0 ]
+then
+    echo "Failed to donwload elmasy.tar!"
+    echo "$OUTPUT"
     exit 1
 fi
 
+# Check if Elmasy installed before, and stop the running service
 if [ -e "/lib/systemd/system/elmasy.service" ]
 then
 
@@ -34,17 +48,39 @@ fi
 
 if [ -e "/opt/elmasy-old" ]
 then
+    echo "Removing elmasy-old..."
     rm -r "/opt/elmasy-old"
 fi
 
-echo "Moving old files to /opt/elmasy-old ..."
-mv /opt/elmasy /opt/elmasy-old
+if [ -e /opt/elmasy ]
+then
+    echo "Moving old files to /opt/elmasy-old ..."
+    mv /opt/elmasy /opt/elmasy-old
+fi
 
 echo "Extracting new files..."
-tar -xf $1 -C /opt
+tar -xf elmasy.tar -C /opt
 
-cp /opt/elmasy-old/elmasy.conf /opt/elmasy
+if [ -e /opt/elmasy-old ]
+then
+    OUTPUT=$(diff /opt/elmasy/elmasy.conf /opt/elmasy-old/elmasy.conf)
+    if [ $? != 0 ]
+    then
+        echo "Changes in the new config:"
+        echo "New < | > old"
+        echo "$OUTPUT"
+    fi
+
+    echo "Copy the old config file to its new place..."
+    cp /opt/elmasy-old/elmasy.conf /opt/elmasy
+fi
+
+echo "Setting executable capabilities..."
+setcap cap_net_admin,cap_net_raw=eip /opt/elmasy/elmasy
+
+echo "installing the new service file..."
 cp /opt/elmasy/elmasy.service /lib/systemd/system/elmasy.service
 systemctl daemon-reload
 
-systemctl start elmasy
+echo "Removing leftover elmasy.tar..."
+rm elmasy.tar
